pyproject.toml
[project]
name = "de-zoomcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clickhouse-sqlalchemy>=0.3.2",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
]
